Implement ThBrk in thbrk.c.
This keeps the imeplementation in one place, and means that brk-common.c
does not need to know about ThBrk.

* src/thbrk/thbrk.c:
* src/thbrk/brk-common.c:
  - Move the implementation of ThBrk from brk-common.c to thbrk.c,
    including the shared instance.

* src/thbrk/thbrk-priv.h:
  - New header file for non-public declarations of thbrk.c.

* src/thbrk/brk-common.h:
  - Export brk_load_default_dict().
  - Remove declarations for accessing ThBrk.

* src/libthai.c:
* src/thbrk/brk-maximal.c:
  - Use thbrk-priv.h instead of brk-common.h.

* src/thbrk/Makefile.am:
  - Add the new header file.
